+++ lib/rust/ensogl/component/selector/src/shape.rs
@@ -195,8 +195,8 @@ pub mod right_overflow {
 
 use enso_frp::Network;
 use ensogl_core::frp::io::Mouse;
+use ensogl_core::gui::component::PointerTarget;
 use ensogl_core::gui::component::ShapeView;
-use ensogl_core::gui::component::ShapeViewEvents;
 
 pub use super::frp::*;
 pub use super::model::*;
@@ -208,7 +208,7 @@ use ensogl_core::display::Scene;
 /// Dragging is ended by a mouse up.
 pub fn shape_is_dragged(
     network: &Network,
-    shape: &ShapeViewEvents,
+    shape: &PointerTarget,
     mouse: &Mouse,
 ) -> enso_frp::Stream<bool> {
     enso_frp::extend! { network
@@ -259,7 +259,7 @@ mod tests {
     fn test_shape_is_dragged() {
         let network = enso_frp::Network::new("TestNetwork");
         let mouse = enso_frp::io::Mouse::default();
-        let shape = ShapeViewEvents::default();
+        let shape = PointerTarget::default();
 
         let is_dragged = shape_is_dragged(&network, &shape, &mouse);
         let _watch = is_dragged.register_watch();
